chore: move the Allowance type definition to the correct place, and adjust fields
