+++ packages/deployments/src/fetchRolesMod.ts
@@ -1,5 +1,11 @@
 import { chains } from "./chains"
-import { ChainId, Clearance, ExecutionOptions, Function } from "./types"
+import {
+  Allowance,
+  ChainId,
+  Clearance,
+  ExecutionOptions,
+  Function,
+} from "./types"
 
 type Props = {
   address: `0x${string}`
@@ -118,15 +124,6 @@ export interface RoleSummary {
   targets: TargetSummary[]
 }
 
-export interface Allowance {
-  key: `0x${string}`
-  refill: bigint
-  maxRefill: bigint
-  period: number
-  balance: bigint
-  timestamp: number
-}
-
 export interface RolesModifier {
   address: `0x${string}`
   owner: `0x${string}`
@@ -144,9 +141,9 @@ const mapGraphQl = (rolesModifier: any): RolesModifier => ({
     key: allowance.key,
     refill: BigInt(allowance.refill),
     maxRefill: BigInt(allowance.maxRefill),
-    period: Number(allowance.period),
+    period: BigInt(allowance.period),
     balance: BigInt(allowance.balance),
-    timestamp: Number(allowance.timestamp),
+    timestamp: BigInt(allowance.timestamp),
   })),
   multiSendAddresses: rolesModifier.unwrapAdapters.map(
     (adapter: any) => adapter.targetAddress
